grup: repurpose OSTREE_GRUB2_EXEC envvar
Originally OSTREE_GRUB2_EXEC was to provide a path to
the grub executible (for initial deployment), but with
the new implementation this is not necessary anymore.
In ostree-grub-generator case path in deployment always
is /usr/lib/ostree/, in system grub2-mkconfig path can
be configured when building ostree.

We can still reuse the OSTREE_GRUB2_EXEC={system,builtin}
for auto test purposes. This envvar can be used to
overwrite compile time decision.