ostree-union-fuse: new fuse filesystem to access multiple commits
it allows to union mount several commits, it is useful with Docker
imported images that are made of several layers, as:

ostree-union-fuse -o whiteouts -o repo=/ostree/repo/ -o layers=layer-1:layer-2 mnt

Currently "atomic mount" is using "checkout --union" for mounting an
image stored in OSTree.  While this is the ideal solution for checking
out an image or in general files to the file system, it has some
problems to keep the transient mount/unmount semantic that atomic has
with Docker images.  To workaround these problems, we were using xattr's
or a text file simulating those when xattr's are not available.

ostree-union-fuse has a series of improvements over it:

- no files are checked out to the file system.
- the mount appears in /proc/mounts.
- umount can be used too (not only atomic umount).
- on a crash/reboot no files are left on the system.
- we don't need xattr's to store the settings of the mount.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>