commit: Support merging docker-style layers
This adds support to OstreeMutableTree for merging one
mtree into another, using the mechanism for whiteouts that
docker/OCI uses (i.e. .wh. whiteout files).

Since the spec doesn't guarantee the order of the whiteouts
(i.e. a file from this layer can appear before a whiteout
that affects it) we can't really apply layer in a streaming
fashion. Thus we read the layer into an mtree end then we
apply it.

This adds two operations to commit:
 --tree=layer-tar=TARFILE
 and
 --tree=layer-ref=COMMIT

which are different ways to apply layers when commiting.