Add OstreeJson base class for json (de)marshalling
This is a baseclass for json objects which can be used to marshal
and demarshall the OCI json formats.

OCI use all kinds of formats for the property names (mediaType,
MemorySwap, diff_ids, os.version, etc) which makes the normal
json-glib gproperty-based marshalling not work (due to property name
canonicalization), and it has some other weird things, so we need a
custom version.